chatapp/rag-chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
